# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ServiceFlow SDM API
  version: 1.0.0
  description: |
    Service Delivery Manager backend API for MGM service delivery operations.

    ## Authentication
    All protected endpoints require a Bearer JWT access token obtained from
    `POST /api/v1/auth/login` or `POST /api/v1/auth/sso/exchange`.

    ## Roles
    Roles in order of privilege: `admin > sdm > manager > tam > csm > engineer > viewer`

    `viewer` has read-only access scoped to their assigned properties/incidents.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Property:
      type: object
      required: [property_id, name, property_type]
      properties:
        property_id:
          type: string
          format: uuid
        name:
          type: string
          example: Bellagio
        property_type:
          type: string
          enum: [hotel, casino, resort, venue]
        created_at:
          type: string
          format: date-time
...
    Incident:
      type: object
      required: [incident_id, incident_number, property_id, title, priority, status]
      properties:
        incident_id:
          type: string
          format: uuid
        incident_number:
          type: integer
        property_id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        priority:
          type: string
          enum: [P1, P2, P3, P4]
        status:
          type: string
          enum: [open, acknowledged, investigating, in-progress, pending, resolved, closed, cancelled]
        tac_case_number:
          type: string
          nullable: true
        reported_by:
          type: string
          format: uuid
        assigned_to:
          type: string
          format: uuid
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    IncidentUpdate:
      type: object
      required: [update_id, incident_id, user_id, content, created_at]
      properties:
        update_id:
          type: string
          format: uuid
        incident_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
...
    UserPropertyAssignment:
      type: object
      required: [property_id, name, property_type]
      properties:
        property_id:
          type: string
          format: uuid
        name:
          type: string
        property_type:
          type: string
          enum: [hotel, casino, resort, venue]
paths:
...
  /properties:
    get:
      summary: List all properties
      tags: [Properties]
      responses:
        "200":
          description: Array of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create a property
      description: Requires role `admin`, `sdm`, or `manager`.
      tags: [Properties]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, propertyType]
              properties:
                name:
                  type: string
                propertyType:
                  type: string
                  enum: [hotel, casino, resort, venue]
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /properties/{propertyId}:
    get:
      summary: Get a single property
      tags: [Properties]
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "200":
          description: Property object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Devices ────────────────────────────────────────────────────────────────
...
  /incidents:
    get:
      summary: List incidents
      description: |
        Privileged roles see all incidents. `viewer` role sees only incidents
        they reported or are assigned to.
      tags: [Incidents]
      responses:
        "200":
          description: Array of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      summary: Create an incident
      description: Requires any privileged role (not viewer).
      tags: [Incidents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [propertyId, title, description, priority]
              properties:
                propertyId:
                  type: string
                  format: uuid
                deviceId:
                  type: string
                  format: uuid
                title:
                  type: string
                  minLength: 5
                description:
                  type: string
                  minLength: 5
                priority:
                  type: string
                  enum: [P1, P2, P3, P4]
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /incidents/{incidentId}:
    get:
      summary: Get a single incident
      description: Viewers receive 404 if they are not the reporter or assignee.
      tags: [Incidents]
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "200":
          description: Incident detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update incident status
      description: Requires role `admin`, `sdm`, `tam`, `engineer`, or `manager`.
      tags: [Incidents]
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [open, acknowledged, investigating, in-progress, pending, resolved, closed, cancelled]
      responses:
        "200":
          description: Updated incident status
          content:
            application/json:
              schema:
                type: object
                properties:
                  incident_id:
                    type: string
                    format: uuid
                  status:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /incidents/{incidentId}/updates:
    post:
      summary: Add an update note to an incident
      tags: [Incidents]
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Update created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentUpdate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /incidents/{incidentId}/tac-link:
    post:
      summary: Link a TAC case to an incident
      description: Requires role `admin`, `sdm`, `tam`, or `engineer`.
      tags: [Incidents]
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tacCaseNumber, tacSeverity]
              properties:
                tacCaseNumber:
                  type: string
                  minLength: 3
                tacSeverity:
                  type: string
                  enum: ["1", "2", "3", "4"]
      responses:
        "200":
          description: TAC link created
          content:
            application/json:
              schema:
                type: object
                properties:
                  incident_id:
                    type: string
                    format: uuid
                  tac_case_number:
                    type: string
                  tac_severity:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── TAC Cases ──────────────────────────────────────────────────────────────
...
  /admin/users/{userId}/properties:
    get:
      summary: List properties assigned to a user
      description: Requires role `admin`, `sdm`, or `manager`.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Array of property assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPropertyAssignment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      summary: Assign properties to a user (additive)
      description: Requires role `admin`, `sdm`, or `manager`.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [propertyIds]
              properties:
                propertyIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
      responses:
        "201":
          description: Properties assigned
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    format: uuid
                  assigned:
                    type: array
                    items:
                      type: string
                      format: uuid
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    put:
      summary: Replace all property assignments for a user
      description: Requires role `admin` only. Deletes existing assignments and sets new ones.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [propertyIds]
              properties:
                propertyIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
      responses:
        "200":
          description: Properties replaced
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    format: uuid
                  properties:
                    type: array
                    items:
                      type: string
                      format: uuid
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{userId}/properties/{propertyId}:
    delete:
      summary: Remove one property assignment from a user
      description: Requires role `admin`, `sdm`, or `manager`.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Assignment removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Shared parameters & responses ──────────────────────────────────────────
...